♻️ Refactor: fetchUserInfo 로직 useHook으로 분리
